# v18 test runs/build-b-codex/contracts/openapi/previous.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Project Scaffolding + Prisma Schema API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/projects/{projectId}/tasks": {
      "get": {
        "operationId": "findAllByProject",
        "tags": [
          "api"
        ],
        "responses": {
          "200": {
            "description": "Generated response"
          }
        },
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "priority",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assignee_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "post": {
        "operationId": "create",
        "tags": [
          "api"
        ],
        "responses": {
          "201": {
            "description": "Generated response"
          }
        },
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string",
                    "enum": [
                      "string",
                      "null"
                    ]
                  },
                  "priority": {
                    "$ref": "#/components/schemas/TaskPriority"
                  },
                  "due_date": {
                    "type": "string",
                    "enum": [
                      "string",
                      "null"
                    ]
                  },
                  "assignee_id": {
                    "type": "string",
                    "enum": [
                      "string",
                      "null"
                    ]
                  }
                },
                "required": [
                  "title",
                  "description",
                  "priority",
                  "due_date",
                  "assignee_id"
                ]
              }
            }
          }
        }
      }
    },
    "/api/tasks/{id}": {
      "get": {
        "operationId": "findOne",
        "tags": [
          "api"
        ],
        "responses": {
          "200": {
            "description": "Generated response"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "patch": {
        "operationId": "update",
        "tags": [
          "api"
        ],
        "responses": {
          "200": {
            "description": "Generated response"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "priority": {
                    "$ref": "#/components/schemas/TaskPriorityEnum"
                  },
                  "due_date": {
                    "type": "string"
                  },
                  "assignee_id": {
                    "type": "string"
                  }
                },
                "required": [
                  "title",
                  "description",
                  "priority",
                  "due_date",
                  "assignee_id"
                ]
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "remove",
        "tags": [
          "api"
        ],
        "responses": {
          "204": {
            "description": "Generated response"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/tasks/{id}/status": {
      "patch": {
        "operationId": "transitionStatus",
        "tags": [
          "api"
        ],
        "responses": {
          "200": {
            "description": "Generated response"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "$ref": "#/components/schemas/TaskStatus"
                  }
                },
                "required": [
                  "status"
                ]
              }
            }
          }
        }
      }
    },
...
    "/projects/{projectId}/tasks": {
      "get": {
        "operationId": "findAllByProject",
        "tags": [
          "projects"
        ],
        "responses": {
          "200": {
            "description": "Generated response"
          }
        },
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "priority",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assignee_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "post": {
        "operationId": "create",
        "tags": [
          "projects"
        ],
        "responses": {
          "201": {
            "description": "Generated response"
          }
        },
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string",
                    "enum": [
                      "string",
                      "null"
                    ]
                  },
                  "priority": {
                    "$ref": "#/components/schemas/TaskPriority"
                  },
                  "due_date": {
                    "type": "string",
                    "enum": [
                      "string",
                      "null"
                    ]
                  },
                  "assignee_id": {
                    "type": "string",
                    "enum": [
                      "string",
                      "null"
                    ]
                  }
                },
                "required": [
                  "title",
                  "description",
                  "priority",
                  "due_date",
                  "assignee_id"
                ]
              }
            }
          }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "operationId": "findOne",
        "tags": [
          "tasks"
        ],
        "responses": {
          "200": {
            "description": "Generated response"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "patch": {
        "operationId": "update",
        "tags": [
          "tasks"
        ],
        "responses": {
          "200": {
            "description": "Generated response"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "priority": {
                    "$ref": "#/components/schemas/TaskPriorityEnum"
                  },
                  "due_date": {
                    "type": "string"
                  },
                  "assignee_id": {
                    "type": "string"
                  }
                },
                "required": [
                  "title",
                  "description",
                  "priority",
                  "due_date",
                  "assignee_id"
                ]
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "remove",
        "tags": [
          "tasks"
        ],
        "responses": {
          "204": {
            "description": "Generated response"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/tasks/{id}/status": {
      "patch": {
        "operationId": "transitionStatus",
        "tags": [
          "tasks"
        ],
        "responses": {
          "200": {
            "description": "Generated response"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "$ref": "#/components/schemas/TaskStatus"
                  }
                },
                "required": [
                  "status"
                ]
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TaskStatus": {
        "type": "string",
        "enum": [
          "TODO",
          "IN_PROGRESS",
          "IN_REVIEW",
          "DONE"
        ]
      },
      "TaskPriority": {
        "type": "string",
        "enum": [
          "LOW",
          "MEDIUM",
          "HIGH",
          "URGENT"
        ]
      },
...
      "TaskPriorityEnum": {
        "type": "string",
        "enum": [
          "LOW",
          "MEDIUM",
          "HIGH",
          "URGENT"
        ]
      },
...
    }
  }
}